Added new Humanitarian Aid secondary objective
This mission costs 5 intel by default, selects a random friendly town, and requests the players bring 2 supply crates to within 50m of the town center. Once the supplies are in place, it deletes those supply crates and rewards the player with +5% to civilian reputation (default).

The intel cost and % chance in civilian reputation are configurable in the kp_liberation_config. The mission will also detect if there are currently no friendly towns, display a notification, and refund the correct intel cost.