prisoners now support ACE Handcuffs
Fixed so that when a captive ACE prisoner is brought to an FOB, it is treated as if the prisoner has been retrieved. It's strange that this wasn't the case for a long time. Prisoners should be tied up.
Also, if a tied up prisoner is released, they will immediately escape.
If they can't find a sector to escape to, they will escape to the northeast. (Random is fine, but it's not worth the effort.)