Improve Multi MSP
Added backward compatibility for multi-MSP.
You can now use either the string or array version of KPLIB_b_mobileRespawn.
You can continue to use presets from last month.

If "KPLIB_b_mobileRespawn" is a string, mobile respawn will not be abbreviated to MSP as before, but will use the stringtable characters.

Changed the "Custom" preset to a sample that supports multi-MSP. Please use this as a reference.

Changed the number of classes that can be specified for MSP from 3 to infinite.
You can add more than 10 vehicles to MSP if necessary.